requirements.txt
pillow==10.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
aiohttp==3.9.1
exif==1.6.0
tqdm==4.66.1
lxml==5.0.0
anthropic==0.8.1
pandas==2.2.1 